rm duplicated test code
