+++ src/test/groovy/com/spidasoftware/mongodb/data/MongoDBFeatureSourceSpec.groovy
@@ -1,13 +1,12 @@
 package com.spidasoftware.mongodb.data
 
 import com.mongodb.BasicDBList
-import com.mongodb.BasicDBObject
 import com.mongodb.MongoClientSettings
 import com.mongodb.ServerAddress
 import com.mongodb.client.MongoClient
 import com.mongodb.client.MongoClients
 import com.mongodb.client.MongoDatabase
-import groovy.json.JsonSlurper
+import com.spidasoftware.mongodb.TestJsonUtils
 import org.bson.Document
 import com.spidasoftware.mongodb.feature.collection.MongoDBFeatureCollection
 import org.geotools.data.Query
@@ -33,47 +32,8 @@ class MongoDBFeatureSourceSpec extends Specification {
     @Shared BasicDBList jsonMapping
     @Shared String namespace = "http://spida/db"
 
-    private static BasicDBList parseJsonResource(String resourcePath) {
-        def jsonSlurper = new JsonSlurper()
-        def parsed = jsonSlurper.parse(MongoDBFeatureSourceSpec.class.getResourceAsStream(resourcePath))
-        return convertToBasicDBList(parsed)
-    }
-
-    private static Document parseJsonResourceAsDocument(String resourcePath) {
-        def jsonSlurper = new JsonSlurper()
-        def parsed = jsonSlurper.parse(MongoDBFeatureSourceSpec.class.getResourceAsStream(resourcePath))
-        return new Document(convertToBasicDBObject(parsed))
-    }
-
-    private static Object convertValue(Object value) {
-        if (value instanceof BigDecimal) {
-            return ((BigDecimal) value).doubleValue()
-        } else if (value instanceof Map) {
-            return convertToBasicDBObject(value)
-        } else if (value instanceof List) {
-            return convertToBasicDBList(value)
-        }
-        return value
-    }
-
-    private static BasicDBList convertToBasicDBList(List list) {
-        BasicDBList dbList = new BasicDBList()
-        list.each { item ->
-            dbList.add(convertValue(item))
-        }
-        return dbList
-    }
-
-    private static BasicDBObject convertToBasicDBObject(Map map) {
-        BasicDBObject dbObject = new BasicDBObject()
-        map.each { key, value ->
-            dbObject.put(key, convertValue(value))
-        }
-        return dbObject
-    }
-
     void setupSpec() {
-        locationJSON = parseJsonResourceAsDocument('/location.json')
+        locationJSON = TestJsonUtils.parseJsonResourceAsDocument(MongoDBFeatureSourceSpec.class, '/location.json')
         String host = System.getProperty("mongoHost")
         String port = System.getProperty("mongoPort")
         String databaseName = System.getProperty("mongoDatabase")
@@ -82,7 +42,7 @@ class MongoDBFeatureSourceSpec extends Specification {
             .applyToClusterSettings { builder -> builder.hosts([serverAddress]) }
             .build()
         MongoClient mongoClient = MongoClients.create(settings)
-        jsonMapping = parseJsonResource('/mapping.json')
+        jsonMapping = TestJsonUtils.parseJsonResource(MongoDBFeatureSourceSpec.class, '/mapping.json')
         mongoDBDataAccess = new MongoDBDataAccess(namespace, host, port, databaseName, null, null, null, jsonMapping)
         database = mongoClient.getDatabase(databaseName)
         database.getCollection("locations").deleteOne(new Document("id", locationJSON.get("id")))
